Fix the two real defects the unmuted warnings found
Environment.cpp:49 printed a size_t through %d. Genuine undefined behaviour
per the standard, though it happens to work on x86-64 where varargs pass in
registers wide enough to hide it. Now %zu with an explicit cast. The
function is diagnostic-only, so impact was low.

C_LZP/C_LZP.cpp:144 wrote `a || b && c` without parentheses. The precedence
was already correct -- && binds tighter, which is what the code intends --
so this is a readability fix with no behavioural change, not a bug fix.

CompressionLibrary.cpp:442 is documented rather than changed. Calling
through a function pointer of a different type is UB by the letter of the
standard, and it is also how every dynamic symbol lookup works: LoadFromDLL
returns a generic FARPROC and only the call site knows the real signature.
Laundering the cast through void* would silence the diagnostic without
removing the UB, which is worse than leaving it visible, so a comment
explains why it stands.

Not fixed, deliberately: 168 of the 239 warnings come from vendored code
(LibTomCrypt ciphers/hashes/prngs, the LZMA and 7z SDKs, zstd, libbsc).
Fixing those means forking upstream.

Of the 68 warnings in DArc-owned code the remainder are cosmetic --
31 -Wwritable-strings, 15 -Wformat-zero-length, 10 -Wunused-parameter --
and worth a separate sweep rather than mixing into this one.

Two warnings that looked alarming and are not, checked rather than assumed:
the /dev/urandom read at Environment.cpp:583 compares size_t against int,
but file_read is fread, which cannot return negative, so the entropy check
cannot silently accept a short read. The bounds checks at
CompressionLibrary.cpp:521 and :540 compare a monotonic counter that starts
at zero against a table size, so the signed-to-unsigned conversion cannot
wrap.

+++ Environment.cpp
@@ -46,7 +46,7 @@ LargestMemoryBlock::LargestMemoryBlock() : p(NULL)
 void LargestMemoryBlock::test()
 {
   if ((size>>20)>0) {
-    printf("Allocated %4d mb, addr=%p\n", size>>20, p);
+    printf("Allocated %4zu mb, addr=%p\n", (size_t)(size>>20), p);
     LargestMemoryBlock next;
     next.test();
   } else {
